feat: search engine with RxJS; move it to handle service (type: string and array)
