+++ src/app/components-ui/table/multiple-checkbox-filter/components/data-table/data-table.component.ts
@@ -1,7 +1,8 @@
 import { Component, OnInit, Self } from '@angular/core';
 import { PageEvent } from '@angular/material/paginator';
 import { TableColumn } from '@lib/components/custom-table/custom-table.component';
-import { BehaviorSubject, catchError, combineLatest, map, Observable, startWith, Subject, switchMap, throwError } from 'rxjs';
+import { FilterSchema, HandleService } from '@lib/services/base/handle.service';
+import { BehaviorSubject, catchError, map, Observable, startWith, Subject, switchMap, throwError } from 'rxjs';
 import { Filter } from '../../models/filter.model';
 import { GithubApi, GithubIssue } from '../../models/service.model';
 import { GithubRepoIssuesService } from '../../service/github-repo-issues.service';
@@ -33,6 +34,7 @@ export class DataTableComponent implements OnInit {
   private _pageIndex$ = new Subject<number>();
 
   constructor(
+    private readonly handle: HandleService,
     @Self() readonly service: GithubRepoIssuesService
   ) { }
 
@@ -54,25 +56,18 @@ export class DataTableComponent implements OnInit {
       }),
     );
 
-    this.issues$ = combineLatest([data$, this.filters$]).pipe(
-      map(([data, params]) =>
-        data.filter((item: GithubIssue) => {
-          let conditions = true;
-          const filterValues = JSON.parse(JSON.stringify(params));
+    const schemas = {
+      query: {
+        type: 'string',
+        enums: ['number', 'title']
+      },
+      state: {
+        type: 'array',
+        enums: ['open', 'closed']
+      }
+    } as FilterSchema;
 
-          for (const key in filterValues) {
-            if (key === 'query') {
-              const searchTerm = item.number + item.title;
-              conditions = conditions && searchTerm.toLowerCase().indexOf(filterValues['query'].trim().toLowerCase()) !== -1;
-            }
-            else if (filterValues[key] !== null && filterValues[key].length) {
-              conditions = conditions && filterValues[key].includes(item[key].trim().toLowerCase());
-            }
-          }
-
-          return conditions;
-        })
-      ),
+    this.issues$ = this.handle.filteredData(data$, this.filters$, schemas).pipe(
       catchError(({ message }) => {
         this.errorMessage$.next(message);
         return throwError(() => new Error(message));
